Collapsing LoadEnsuring and Staleable and other unnecessary modules into Node
(cherry picked from commit 80d32dc7bbe8a1bf700383d18d304e316e17841c)
