--- lib/ae_page_objects/concerns/load_ensuring.rb
@@ -1,26 +1,0 @@
-module AePageObjects
-  module Concerns
-    module LoadEnsuring
-
-      def initialize(*args)
-        super
-        ensure_loaded!
-      end
-
-    private
-
-      def loaded_locator
-      end
-
-      def ensure_loaded!
-        if locator = loaded_locator
-          find(*eval_locator(locator))
-        end
-
-        self
-      rescue Capybara::ElementNotFound => e
-        raise LoadingElementFailed, e.message
-      end
-    end
-  end
-end
